refactor(spendings): extract FilterCategory type (COS-103)
Move the FilterCategory interface out of SpendingCategoryFilter into a
dedicated spendings/interfaces/spendingCategoryFilterTypes.ts, matching
the existing spendingCategoryBreakdownTypes.ts pattern. Pure type move,
no runtime change.
