+++ front/src/components/spendings/view/SpendingView.tsx
@@ -26,9 +26,9 @@ import { Plus } from "lucide-react";
 import { useEffect, useMemo, useState } from "react";
 
 import type { BreakdownRow } from "@components/spendings/interfaces/spendingCategoryBreakdownTypes";
+import type { FilterCategory } from "@components/spendings/interfaces/spendingCategoryFilterTypes";
 import type { MonthRange } from "@components/spendings/interfaces/spendingDashboardTypes";
 import type { SpendingDayGroup } from "@components/spendings/types";
-import type { FilterCategory } from "@components/spendings/view/SpendingCategoryFilter";
 
 const FALLBACK_COLOR = CATEGORY_FALLBACK;
 const UNCATEGORIZED_KEY = "none";
